requirements.txt
langchain==0.3.27
langchain_chroma==0.2.5
langchain_community==0.3.29
langchain_core==0.3.75
langchain_huggingface==0.3.1
langchain_openai==0.3.32
langchain_text_splitters==0.3.9
python-dotenv==1.1.1
streamlit==1.49.0
sentence-transformers==2.2.2
torch==2.0.1
pypdf==3.7.3
python-dotenv==1.0.0
openai==0.27.0